converting model and input data to half precision

+++ animate.py
@@ -29,7 +29,7 @@ def normalize_kp(kp_source, kp_driving, kp_driving_initial, adapt_movement_scale
         kp_new['value'] = kp_value_diff + kp_source['value']
 
         if use_relative_jacobian:
-            jacobian_diff = torch.matmul(kp_driving['jacobian'], torch.inverse(kp_driving_initial['jacobian']))
+            jacobian_diff = torch.matmul(kp_driving['jacobian'], torch.inverse(kp_driving_initial['jacobian'].float()).half())
             kp_new['jacobian'] = torch.matmul(jacobian_diff, kp_source['jacobian'])
 
     return kp_new
